Feature al

+++ src/main/java/group5/sebm/controller/dto/UserDto.java
@@ -7,12 +7,12 @@ import lombok.Setter;
 @Getter
 public class UserDto {
 
-  private Integer id;
+  private Long id;
   private String username;
   private String password;
   private Integer age;
 
-  public UserDto(Integer id, String username, String password, Integer age) {
+  public UserDto(Long id, String username, String password, Integer age) {
     this.id = id;
     this.username = username;
     this.password = password;
